refactor: moved other editor related modals to editor modals folder
